refactor: fix edit task modal naming typo
