+++ src/components/TasksList/index.tsx
@@ -2,7 +2,7 @@ import { Button, Grid } from "@mui/material"
 import { useApp } from "../../contexts/AppContext"
 import { Task, useTasks } from "../../contexts/TasksContext"
 import CreateTaskModal from "../CreateTaskModal"
-import EditTaskModal from "../EdtiTaskModal"
+import EditTaskModal from "../EditTaskModal"
 import FloatingAddButton from "../FloatingAddButton"
 import TaskDetailsModal from "../TaskDetailsModal"
 import TaskItem from "../TaskItem"
